# storage/api-docs/api-docs.json
{
    "swagger": "2.0",
    "info": {
        "title": "API Documentation BNN",
        "version": "1.0.0"
    },
    "basePath": "/bnn_api/public/api",
    "paths": {
...
        "/login": {
            "post": {
                "tags": [
                    "Login"
                ],
                "summary": "Login user",
                "operationId": "getToken",
                "parameters": [
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Email user.",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "Password user.",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
        "/users": {
            "get": {
                "tags": [
                    "User"
                ],
                "summary": "Get List User",
                "operationId": "get data",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            },
            "post": {
                "tags": [
                    "User"
                ],
                "summary": "Create User",
                "operationId": "postUser",
                "parameters": [
                    {
                        "name": "group_id",
                        "in": "formData",
                        "description": "ID Group",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "user_name",
                        "in": "formData",
                        "description": "username",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user_login",
                        "in": "formData",
                        "description": "user login",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user_email",
                        "in": "formData",
                        "description": "User email",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user_pass",
                        "in": "formData",
                        "description": "user password",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "active_flag",
                        "in": "formData",
                        "description": "active flag",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "description",
                        "in": "formData",
                        "description": "deskripsi",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "wilayah_id",
                        "in": "formData",
                        "description": "id wilayah",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "user_pass_decrypt",
                        "in": "formData",
                        "description": "password decrypt",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "remember_token",
                        "in": "formData",
                        "description": "token",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "email",
                        "in": "formData",
                        "description": "Email user",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "password",
                        "in": "formData",
                        "description": "password",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "nip",
                        "in": "formData",
                        "description": "NIP",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "flag",
                        "in": "formData",
                        "description": "ID Group",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "jabatan_pegawai",
                        "in": "formData",
                        "description": "jabatan pegawai",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "telp_pegawai",
                        "in": "formData",
                        "description": "no telepon pegawai",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "lokasi_kerja",
                        "in": "formData",
                        "description": "Lokasi kerja",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "foto_pegawai",
                        "in": "formData",
                        "description": "Foto Pegawai",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "email_token",
                        "in": "formData",
                        "description": "token email",
                        "required": false,
                        "type": "string"
                    },
                    {
                        "name": "temp_change_email",
                        "in": "formData",
                        "description": "ID Group",
                        "required": false,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "406": {
                        "description": "not acceptable"
                    },
                    "500": {
                        "description": "internal server error"
                    }
                }
            }
        },
...
    },
...
}